Add test cases for DMDA with cusp vectors.
The single process test case passes but it is not valgrind clean. The
scatter types used in VecView_MPI_DA are not initialized properly.
Apparently this doesn't cause problems in the single processor case but
I think the failures seen in the multi-process cases are related to
this. So this needs to be fixed.

The two process test case segfaults (presumably because of badly
initialized scatters).